table editor works

+++ src/gui.c
@@ -29,6 +29,7 @@
 #include "editor_gadget.h"
 #include "settings_gadget.h"
 #include "image_editor.h"
+#include "table_editor.h"
 
 
 #include "debugging_utils.h"
@@ -62,7 +63,7 @@ enum
 
 static void FreeGUIObjects (APTR app_p);
 
-static APTR CreateGUIObjects (struct MUI_CustomClass *editor_class_p, struct MUI_CustomClass *prefs_class_p, struct MUI_CustomClass *image_editor_class_p, MDPrefs *prefs_p);
+static APTR CreateGUIObjects (struct MUI_CustomClass *editor_class_p, struct MUI_CustomClass *settings_class_p, struct MUI_CustomClass *image_editor_class_p, struct MUI_CustomClass *table_editor_class_p, MDPrefs *prefs_p);
 
 static void RunMD (APTR app_p);
 
@@ -129,48 +130,63 @@ BOOL CreateMUIInterface (MDPrefs *prefs_p)
 
 					if (image_editor_class_p)
 						{
-							APTR app_p;
+							struct MUI_CustomClass *table_editor_class_p;
 
 							DB (KPRINTF ("%s %ld - Inited Image Editor\n", __FILE__, __LINE__));
 
-							app_p = CreateGUIObjects (editor_class_p, settings_class_p, image_editor_class_p, prefs_p);
+							table_editor_class_p = InitTableEditorClass ();
 
-							if (app_p)
+							if (table_editor_class_p)
 								{
-									CONST CONST_STRPTR md_reg_s = "#?.md";
-									const size_t md_reg_length = strlen (md_reg_s);
-									const size_t size = (md_reg_length + 1) << 1;
-
-									DB (KPRINTF ("%s %ld - Created GUI Objects\n", __FILE__, __LINE__));
-
-									s_file_pattern_s = (STRPTR) IExec -> AllocVecTags (size, TAG_DONE);
-
-									if (s_file_pattern_s)
+									APTR app_p;
+		
+									DB (KPRINTF ("%s %ld - Inited Image Editor\n", __FILE__, __LINE__));
+	
+									app_p = CreateGUIObjects (editor_class_p, settings_class_p, image_editor_class_p, table_editor_class_p, prefs_p);
+		
+									if (app_p)
 										{
-											DB (KPRINTF ("%s %ld - Created File Pattern\n", __FILE__, __LINE__));
-
-											if (IDOS -> ParsePattern (md_reg_s, s_file_pattern_s, size) >= 0)
+											CONST CONST_STRPTR md_reg_s = "#?.md";
+											const size_t md_reg_length = strlen (md_reg_s);
+											const size_t size = (md_reg_length + 1) << 1;
+		
+											DB (KPRINTF ("%s %ld - Created GUI Objects\n", __FILE__, __LINE__));
+		
+											s_file_pattern_s = (STRPTR) IExec -> AllocVecTags (size, TAG_DONE);
+		
+											if (s_file_pattern_s)
 												{
-													DB (KPRINTF ("%s %ld - Parsed File Pattern\n", __FILE__, __LINE__));
-
-													RunMD (app_p);
-													success_flag = TRUE;
-
-													/*
-													** save the current weights of all Balance objects until the next reboot
-													** if the weights are to be saved permanently the MUIV_Application_Save_ENVARC must be used instead
-													*/
-													IIntuition -> IDoMethod (app_p, MUIM_Application_Save, MUIV_Application_Save_ENV);
-
-													IExec -> FreeVec (s_file_pattern_s);
+													DB (KPRINTF ("%s %ld - Created File Pattern\n", __FILE__, __LINE__));
+		
+													if (IDOS -> ParsePattern (md_reg_s, s_file_pattern_s, size) >= 0)
+														{
+															DB (KPRINTF ("%s %ld - Parsed File Pattern\n", __FILE__, __LINE__));
+		
+															RunMD (app_p);
+															success_flag = TRUE;
+		
+															/*
+															** save the current weights of all Balance objects until the next reboot
+															** if the weights are to be saved permanently the MUIV_Application_Save_ENVARC must be used instead
+															*/
+															IIntuition -> IDoMethod (app_p, MUIM_Application_Save, MUIV_Application_Save_ENV);
+		
+															IExec -> FreeVec (s_file_pattern_s);
+														}
 												}
-										}
-
-									FreeGUIObjects (app_p);
-								}		/* if (CreateGUIObjects (screen_p, app_port_p, hook_p)) */
+		
+											FreeGUIObjects (app_p);
+										}		/* if (CreateGUIObjects (screen_p, app_port_p, hook_p)) */
+									else
+										{
+											IDOS -> PutStr ("Failed to create the user interface\n");
+										}					
+														
+									FreeTableEditorClass (table_editor_class_p);
+								}
 							else
 								{
-									IDOS -> PutStr ("Failed to create the user interface\n");
+									IDOS -> PutStr ("Failed to set up the table editor class\n");
 								}
 
 							FreeImageEditorClass (image_editor_class_p);
@@ -229,7 +245,7 @@ MakeStaticHook(DroppedFileHook, DroppedFile);
 
 
 
-static APTR CreateGUIObjects (struct MUI_CustomClass *editor_class_p, struct MUI_CustomClass *settings_class_p, struct MUI_CustomClass *image_editor_class_p, MDPrefs *prefs_p)
+static APTR CreateGUIObjects (struct MUI_CustomClass *editor_class_p, struct MUI_CustomClass *settings_class_p, struct MUI_CustomClass *image_editor_class_p, struct MUI_CustomClass *table_editor_class_p, MDPrefs *prefs_p)
 {
 	APTR app_p = NULL;
 	APTR strip_p = NULL;
@@ -238,6 +254,8 @@ static APTR CreateGUIObjects (struct MUI_CustomClass *editor_class_p, struct MUI
 	Object *about_box_p = NULL;
 	Object *image_editor_p = NULL;
 	Object *image_editor_window_p = NULL;
+	Object *table_editor_p = NULL;
+	Object *table_editor_window_p = NULL;
 
 	static const char * const used_classes [] =
 		{
@@ -383,14 +401,26 @@ static APTR CreateGUIObjects (struct MUI_CustomClass *editor_class_p, struct MUI
 				MUIA_Group_Horiz, FALSE,
 
 				MUIA_Group_Child, image_editor_p = IIntuition -> NewObject (image_editor_class_p -> mcc_Class, NULL,
-					ImageButtonFrame,
-					MUIA_FillArea, FALSE,
 					MUIA_ShortHelp, (uint32) "Image Editor",
 				TAG_DONE),
 			TAG_DONE),
 
 		TAG_DONE),
 
+
+		SubWindow, table_editor_window_p = IMUIMaster -> MUI_NewObject (MUIC_Window,
+			MUIA_Window_Title, "Insert Table",
+ 			MUIA_ShortHelp, (uint32) "Edit Table Details",
+
+			WindowContents, IMUIMaster -> MUI_NewObject (MUIC_Group,
+				MUIA_Group_Horiz, FALSE,
+
+				MUIA_Group_Child, table_editor_p = IIntuition -> NewObject (table_editor_class_p -> mcc_Class, NULL,
+					MUIA_ShortHelp, (uint32) "Table Editor",
+				TAG_DONE),
+			TAG_DONE),
+
+		TAG_DONE),
 
 		SubWindow, s_window_p = IMUIMaster -> MUI_NewObject (MUIC_Window,
 			MUIA_Window_Title, "AmiMarkdown",
@@ -527,6 +557,15 @@ static APTR CreateGUIObjects (struct MUI_CustomClass *editor_class_p, struct MUI
 					IIntuition -> SetAttrs (image_editor_p, IEA_Editor, s_editor_p, TAG_DONE);
 				}
 
+			if (table_editor_window_p)
+				{
+					IIntuition -> IDoMethod (toolbar_p, MUIM_TheBar_DoOnButton, BID_TABLE, MUIM_Notify, MUIA_Pressed, FALSE, table_editor_window_p, 3, MUIM_Set, MUIA_Window_Open, TRUE);
+					IIntuition -> IDoMethod (table_editor_window_p, MUIM_Notify, MUIA_Window_CloseRequest, TRUE, MUIV_Notify_Self, 3, MUIM_Set, MUIA_Window_Open, FALSE);
+
+					IIntuition -> SetAttrs (table_editor_p, TEA_Editor, s_editor_p, TAG_DONE);
+				}
+
+
 			if (toolbar_p)
 				{
 		      IIntuition -> IDoMethod (toolbar_p, MUIM_TheBar_DoOnButton, BID_OPEN, MUIM_Notify, MUIA_Pressed, FALSE, s_editor_p, 1, MEM_MDEditor_Load);
@@ -543,8 +582,6 @@ static APTR CreateGUIObjects (struct MUI_CustomClass *editor_class_p, struct MUI
 					IIntuition -> IDoMethod (toolbar_p, MUIM_TheBar_DoOnButton, BID_INDENTED_CODE, MUIM_Notify, MUIA_Pressed, FALSE, s_editor_p, 3, MUIM_Set, MEA_SurroundSelection, MEV_MDEditor_Style_IndentedCode);
 
 					IIntuition -> IDoMethod (toolbar_p, MUIM_TheBar_DoOnButton, BID_HORIZONTAL_RULE, MUIM_Notify, MUIA_Pressed, FALSE, s_editor_p, 3, MUIM_Set, MEA_InsertItem, MEV_MDEditor_HorizontalRule);
-					IIntuition -> IDoMethod (toolbar_p, MUIM_TheBar_DoOnButton, BID_TABLE, MUIM_Notify, MUIA_Pressed, FALSE, s_editor_p, 3, MUIM_Set, MEA_InsertItem, MEV_MDEditor_Table);
-					IIntuition -> IDoMethod (toolbar_p, MUIM_TheBar_DoOnButton, BID_IMAGE, MUIM_Notify, MUIA_Pressed, FALSE, s_editor_p, 3, MUIM_Set, MEA_InsertItem, MEV_MDEditor_Image);
 
 					IIntuition -> IDoMethod (toolbar_p, MUIM_TheBar_DoOnButton, BID_HYPERLINK, MUIM_Notify, MUIA_Pressed, FALSE, s_editor_p, 3, MUIM_Set, MEA_InsertItem, MEV_MDEditor_Hyperlink);
 				}
